+++ packages/components/src/internal/components/entities/actions.ts
@@ -34,7 +34,8 @@ import { getAppHomeFolderPath, getFolderDataExclusion, hasModule } from '../../a
 
 import { resolveErrorMessage } from '../../util/messaging';
 
-import { SAMPLE_MANAGER_APP_PROPERTIES } from '../../app/constants';
+import { SAMPLE_MANAGER_APP_PROPERTIES, SAMPLES_KEY, SOURCES_KEY } from '../../app/constants';
+import { DataTypeRowIdsFromTransactionIds, getGridIdsFromTransactionId } from '../../actions';
 
 import { QueryModel } from '../../../public/QueryModel/QueryModel';
 
@@ -1539,4 +1540,67 @@ export function updateCellValuesForSampleIds(
             resolve({ editorModelChanges, cellKeyChanges });
         }
     });
+}
+
+async function getDataTypesFromTransactionId(
+    transactionAuditId: number | string,
+    auditDataType: string,
+    schemaQuery: SchemaQuery,
+    typeColumn: string
+): Promise<DataTypeRowIdsFromTransactionIds> {
+    if (!transactionAuditId) return { rowIds: [], dataTypeRowCounts: {}, dataTypes: [] };
+
+    const { rowIds, dataTypeRowCounts } = await getGridIdsFromTransactionId(transactionAuditId, auditDataType);
+    const distinct = await selectDistinctRows({
+        schemaName: schemaQuery.schemaName,
+        queryName: schemaQuery.queryName,
+        column: typeColumn,
+        filterArray: [Filter.create('RowId', rowIds, Filter.Types.IN)],
+    });
+    return { rowIds, dataTypeRowCounts, dataTypes: distinct.values };
+}
+
+export function getSampleTypesFromTransactionIds(
+    transactionAuditId: number | string
+): Promise<DataTypeRowIdsFromTransactionIds> {
+    return getDataTypesFromTransactionId(
+        transactionAuditId,
+        SAMPLES_KEY,
+        SCHEMAS.EXP_TABLES.MATERIALS,
+        'SampleSet/Name'
+    );
+}
+
+export async function getDataClassesFromTransactionIds(
+    transactionAuditId: number | string
+): Promise<DataTypeRowIdsFromTransactionIds> {
+    const results = await getDataTypesFromTransactionId(
+        transactionAuditId,
+        SOURCES_KEY,
+        SCHEMAS.EXP_TABLES.DATA,
+        'DataClass/Name'
+    );
+
+    if (!results.rowIds.length) return { ...results, typeNameRowCounts: {} };
+
+    const { dataTypeRowCounts, dataTypes } = results;
+    const dataTypeLcMap = Object.fromEntries((dataTypes ?? []).map(dt => [dt.toLowerCase(), dt]));
+
+    const typeNameRowCounts: Record<string, number> = {};
+    if (Object.keys(dataTypeRowCounts).length > 0) {
+        const dataClasses = await selectRows({
+            schemaQuery: SCHEMAS.EXP_TABLES.DATA_CLASSES,
+            columns: ['Name', 'RowId'],
+            filterArray: [Filter.create('rowId', Object.keys(dataTypeRowCounts), Filter.Types.IN)],
+            containerFilter: Query.containerFilter.currentPlusProjectAndShared,
+        });
+
+        dataClasses.rows.forEach(row => {
+            const dataClassLc = caseInsensitive(row, 'Name')?.value?.toLowerCase();
+            const rowId = caseInsensitive(row, 'RowId').value;
+            typeNameRowCounts[dataTypeLcMap[dataClassLc]] = dataTypeRowCounts[rowId];
+        });
+    }
+
+    return { ...results, typeNameRowCounts };
 }
